Implement basic i386 physical memory allocator
i386 physical memory allocator implements a basic counter where each number
corresponds to a 4096 byte chunk of the Multiboot Memory Map. Allocation
of a page increases the counter by one. Both allocation and lookups are
inefficient due to the need to loop through each chunk of the memory map.
Deallocation is not yet implemented.