Add kernel heap algorithm and memory system abstractions
Implements a heap algorithm that allows allocating and freeing
of memory. Will replace the dumb placement allocator once paging
is enabled.

The heap algorithm deals with contiguous regions of memory called
blocks. Each block contains a kheap_block struct (defined in
kernel/include/mm/heap.h) which contains metadata about the
region and serves as a linked list pointing to the next block.
Each block is split up into individual sections whose size is
defined in the block header. Allocations take place on the
sections themselves, and as such, the minimum amount of memory
that can be allocated is that of the smallest section in the
first block containing enough free memory in the heap. Metadata
containing the availability of each section is stored in a
bitset which is located directly after the kheap_block struct
in each block. Metadata such as the size of the bitset is
stored in the block struct as well. Upon allocation, the linked
list of blocks is traversed until a block containing enough
memory is found. Afterwards, the bitmap is traversed to find
enough contiguous sections to satisfy the allocation request.
The sections are then marked as free in the bitset. In addition,
a kheap_alloc_header struct is placed at the beginning of the
allocated region which contains metadata regarding the size
of the allocation. When the allocation is freed, this size of
the allocation is read from this struct and all corresponding
sections are marked as free in the bitset. The memory is then
zeroed.

By default, the heap contains no blocks. A mechanism to dynamically
map blocks to the heap as required through the paging system will
be developed.

Also add 2 kernel abstraction layers for paging and virtual memory
allocation. The abstractions define interfaces which are implemented
in architecture-specific code and installed at boot. For more
information, see kernel/mm/alloc.c and kernel/mm/paging.c and their
corresponding header files. The abstractions will allow for less
architecture-dependent code in high-level kernel logic and make
porting to future architectures easier.