Improve heap algorithm
Heap algorithm no longer places kheap_alloc_header structs at the
beginning of each allocation (this was previously done to keep track
of allocation sizes for freeing). Instead, a second bitmap is placed
at the beginning of each block after the kheap_block struct and the
original bitmap which is used to keep track of the currently used
sectons. The second bitmap is used to keep track of the last section
in each allocation and is used to determine the length of an allocation
for freeing. While this method has more initial overhead, it is much
more efficient for blocks that contain a high number of allocations.