Add support for page-aligned allocations in kernel heap
To allow for this, the delimiter bitset that belongs to each kheap
block now stores the first section in addition to the last section
of each allocation. This is to allow kheap_free to determine the
starting point of an allocation if a pointer to the middle of the
allocated region was given to it (as is the case with page aligned
memory allocations).